+++ src/components/controls/buttons/SubmitButton.js
@@ -1,9 +1,9 @@
 import { Button } from "@mui/material";
 
 export function SubmitButton(props) {
-  const { onClick } = props;
+  const { onClick, disabled } = props;
   return (
-    <Button variant="contained" onClick={onClick}>
+    <Button disabled={disabled} variant="contained" onClick={onClick}>
       Submit
     </Button>
   );
