+++ src/components/header/Header.js
@@ -1,8 +1,8 @@
 import "./Header.css";
 import { Toolbar, Stepper, StepLabel, Step, Box } from "@mui/material";
 import { useSelector } from "react-redux";
-import { FORM_STEPS } from "../constants";
-import { getCurrentForm } from "../redux/selector";
+import { FORM_STEPS } from "../../constants";
+import { getCurrentForm } from "../../redux/selector";
 
 function Header() {
   const currentForm = useSelector(getCurrentForm);
