+++ src/history/index.js
@@ -1,0 +1,3 @@
+import { createBrowserHistory } from "history";
+
+export default createBrowserHistory();
