+++ src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
+import store from "./redux/store/store";
 import App from "./App";
-import store from "./components/redux/store/store";
 
 const root = document.getElementById("root");
 const rootEle = ReactDOM.createRoot(root);
